Added time based movement. Formatted code base.
I added a command called TimeMove, which takes turn and drive power in
it's constructor as well as the time to run for. I then added this to
AutonomousLeft with AddSequential, and gave it an initial power of
0.1 with a 3 second run time. Note, this command is untested, since
we haven't had time with the robot yet.

I formatted the code base using BSD/Allman Formatting using the Eclipse
Formatter.