Added button push to toggle drive direction.
First cleaned up code in OI class. Made seperate functions for each part
of the robot's buttons and called those functions in the OI constructor.

Then added a new function for drive buttons and added a temporary button
on the right drive joystick for switching drive directions. Will have to
test with the actual joystick in order to get real buttons.

Fixed bugs in the "SetDriveReverse" class. (Inverting the drive
multiplier, and stopping once invertion was done.)

Made drive multiplier a public variable in DriveTrain. It doesn't need
to be private. The options were create a getter and setter or make it
public and making it public has less code.